moved BaseReference into its own module to avoid circular imports
